requirements.txt
python-telegram-bot[fast]==20.3